Working with drawing a square and a rhombus, change the screen size, Ability to display the time using the startTime function
Working with drawing a square and a rhombus (coordinates, sizes, and color); the circle functionality is not yet implemented. Ability to change the screen size by calling the corresponding function from the file (res.asm). Ability to display the time using the startTime function located in the time.asm file (it will be used for recording in the report file).

Upon launching the program:

Drawing a rhombus at the chosen location;
Squares are outlined through a loop.
The entire project consists of the main.asm file and the lib folder.