Added shapes, completed delay. FIX - sequential filling and buffer clearing.
Added:
Drawing of a triangle and a rhombus (coordinates, sizes, and color)
Completed the delay function
Shapes filled with color;
Upon launching the program:
Drawing a square, triangle, rhombus, and circle;
To be completed:
• Random number generator for generating shapes in different positions, with different colors and sizes.
• Response to clicks
P.S. I'm having issues with the circle and the generator; I haven't figured out how to fill the shapes one after another. For this, a method to clear the video buffer is needed (and I don't know it yet...) - after clicking on a shape, the old image is filled, and a new one is drawn.